Handle Neo.ClientError.Security.AuthorizationExpired
 - Close the connection and mark all idle connections in the pool as expired
 - Raise the error to the user or do a retry if inside transaction function

Refactoring shared code of Bolt 3 and Bolt 4 connection into base class.
