+++ tests/unit/io/test_class_bolt4x3.py
@@ -21,7 +21,7 @@
 
 import pytest
 
-from neo4j.io._bolt4 import Bolt4x2
+from neo4j.io._bolt4 import Bolt4x3
 from neo4j.conf import PoolConfig
 
 
@@ -29,7 +29,7 @@ from neo4j.conf import PoolConfig
 def test_conn_is_stale(fake_socket, set_stale):
     address = ("127.0.0.1", 7687)
     max_connection_lifetime = 0
-    connection = Bolt4x2(address, fake_socket(address), max_connection_lifetime)
+    connection = Bolt4x3(address, fake_socket(address), max_connection_lifetime)
     if set_stale:
         connection.set_stale()
     assert connection.stale() is True
@@ -39,7 +39,7 @@ def test_conn_is_stale(fake_socket, set_stale):
 def test_conn_is_not_stale_if_not_enabled(fake_socket, set_stale):
     address = ("127.0.0.1", 7687)
     max_connection_lifetime = -1
-    connection = Bolt4x2(address, fake_socket(address), max_connection_lifetime)
+    connection = Bolt4x3(address, fake_socket(address), max_connection_lifetime)
     if set_stale:
         connection.set_stale()
     assert connection.stale() is set_stale
@@ -49,7 +49,7 @@ def test_conn_is_not_stale_if_not_enabled(fake_socket, set_stale):
 def test_conn_is_not_stale(fake_socket, set_stale):
     address = ("127.0.0.1", 7687)
     max_connection_lifetime = 999999999
-    connection = Bolt4x2(address, fake_socket(address), max_connection_lifetime)
+    connection = Bolt4x3(address, fake_socket(address), max_connection_lifetime)
     if set_stale:
         connection.set_stale()
     assert connection.stale() is set_stale
@@ -58,7 +58,7 @@ def test_conn_is_not_stale(fake_socket, set_stale):
 def test_db_extra_in_begin(fake_socket):
     address = ("127.0.0.1", 7687)
     socket = fake_socket(address)
-    connection = Bolt4x2(address, socket, PoolConfig.max_connection_lifetime)
+    connection = Bolt4x3(address, socket, PoolConfig.max_connection_lifetime)
     connection.begin(db="something")
     connection.send_all()
     tag, fields = socket.pop_message()
@@ -70,7 +70,7 @@ def test_db_extra_in_begin(fake_socket):
 def test_db_extra_in_run(fake_socket):
     address = ("127.0.0.1", 7687)
     socket = fake_socket(address)
-    connection = Bolt4x2(address, socket, PoolConfig.max_connection_lifetime)
+    connection = Bolt4x3(address, socket, PoolConfig.max_connection_lifetime)
     connection.run("", {}, db="something")
     connection.send_all()
     tag, fields = socket.pop_message()
@@ -84,7 +84,7 @@ def test_db_extra_in_run(fake_socket):
 def test_n_extra_in_discard(fake_socket):
     address = ("127.0.0.1", 7687)
     socket = fake_socket(address)
-    connection = Bolt4x2(address, socket, PoolConfig.max_connection_lifetime)
+    connection = Bolt4x3(address, socket, PoolConfig.max_connection_lifetime)
     connection.discard(n=666)
     connection.send_all()
     tag, fields = socket.pop_message()
@@ -103,7 +103,7 @@ def test_n_extra_in_discard(fake_socket):
 def test_qid_extra_in_discard(fake_socket, test_input, expected):
     address = ("127.0.0.1", 7687)
     socket = fake_socket(address)
-    connection = Bolt4x2(address, socket, PoolConfig.max_connection_lifetime)
+    connection = Bolt4x3(address, socket, PoolConfig.max_connection_lifetime)
     connection.discard(qid=test_input)
     connection.send_all()
     tag, fields = socket.pop_message()
@@ -123,7 +123,7 @@ def test_n_and_qid_extras_in_discard(fake_socket, test_input, expected):
     # python -m pytest tests/unit/io/test_class_bolt4x0.py -s -k test_n_and_qid_extras_in_discard
     address = ("127.0.0.1", 7687)
     socket = fake_socket(address)
-    connection = Bolt4x2(address, socket, PoolConfig.max_connection_lifetime)
+    connection = Bolt4x3(address, socket, PoolConfig.max_connection_lifetime)
     connection.discard(n=666, qid=test_input)
     connection.send_all()
     tag, fields = socket.pop_message()
@@ -142,7 +142,7 @@ def test_n_and_qid_extras_in_discard(fake_socket, test_input, expected):
 def test_n_extra_in_pull(fake_socket, test_input, expected):
     address = ("127.0.0.1", 7687)
     socket = fake_socket(address)
-    connection = Bolt4x2(address, socket, PoolConfig.max_connection_lifetime)
+    connection = Bolt4x3(address, socket, PoolConfig.max_connection_lifetime)
     connection.pull(n=test_input)
     connection.send_all()
     tag, fields = socket.pop_message()
@@ -162,7 +162,7 @@ def test_qid_extra_in_pull(fake_socket, test_input, expected):
     # python -m pytest tests/unit/io/test_class_bolt4x0.py -s -k test_qid_extra_in_pull
     address = ("127.0.0.1", 7687)
     socket = fake_socket(address)
-    connection = Bolt4x2(address, socket, PoolConfig.max_connection_lifetime)
+    connection = Bolt4x3(address, socket, PoolConfig.max_connection_lifetime)
     connection.pull(qid=test_input)
     connection.send_all()
     tag, fields = socket.pop_message()
@@ -174,7 +174,7 @@ def test_qid_extra_in_pull(fake_socket, test_input, expected):
 def test_n_and_qid_extras_in_pull(fake_socket):
     address = ("127.0.0.1", 7687)
     socket = fake_socket(address)
-    connection = Bolt4x2(address, socket, PoolConfig.max_connection_lifetime)
+    connection = Bolt4x3(address, socket, PoolConfig.max_connection_lifetime)
     connection.pull(n=666, qid=777)
     connection.send_all()
     tag, fields = socket.pop_message()
@@ -187,7 +187,8 @@ def test_hello_passes_routing_metadata(fake_socket_pair):
     address = ("127.0.0.1", 7687)
     sockets = fake_socket_pair(address)
     sockets.server.send_message(0x70, {"server": "Neo4j/4.2.0"})
-    connection = Bolt4x2(address, sockets.client, PoolConfig.max_connection_lifetime,
+    connection = Bolt4x3(address, sockets.client,
+                         PoolConfig.max_connection_lifetime,
                          routing_context={"foo": "bar"})
     connection.hello()
     tag, fields = sockets.server.pop_message()
